# docs/swagger/swagger.yaml
basePath: /
definitions:
  handler.bookingResponse:
    properties:
      conferenceLink:
        type: string
      createdAt:
        type: string
      id:
        type: string
      slotId:
        type: string
      status:
        type: string
      userId:
        type: string
    type: object
  handler.cancelBookingResponse:
    properties:
      booking:
        $ref: '#/definitions/handler.bookingResponse'
    type: object
  handler.createBookingRequest:
    properties:
      createConferenceLink:
        type: boolean
      slotId:
        type: string
    type: object
  handler.createBookingResponse:
    properties:
      booking:
        $ref: '#/definitions/handler.bookingResponse'
    type: object
  handler.createRoomRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  handler.createRoomResponse:
    properties:
      room:
        $ref: '#/definitions/handler.roomResponse'
    type: object
  handler.createScheduleRequest:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      startTime:
        type: string
    type: object
  handler.createScheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/handler.scheduleResponse'
    type: object
  handler.dummyLoginRequest:
    properties:
      role:
        type: string
    type: object
  handler.listAllBookingsResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/handler.bookingResponse'
        type: array
      pagination:
        $ref: '#/definitions/handler.paginationResponse'
    type: object
  handler.listBookingsResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/handler.bookingResponse'
        type: array
    type: object
  handler.listRoomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/handler.roomResponse'
        type: array
    type: object
  handler.listSlotsResponse:
    properties:
      slots:
        items:
          $ref: '#/definitions/handler.slotResponse'
        type: array
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.paginationResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  handler.registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  handler.registerResponse:
    properties:
      user:
        $ref: '#/definitions/handler.userResponse'
    type: object
  handler.roomResponse:
    properties:
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  handler.scheduleResponse:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      id:
        type: string
      roomId:
        type: string
      startTime:
        type: string
    type: object
  handler.slotResponse:
    properties:
      end:
        type: string
      id:
        type: string
      roomId:
        type: string
      start:
        type: string
    type: object
  handler.tokenResponse:
    properties:
      token:
        type: string
    type: object
  handler.userResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: API for room booking service.
  title: Room Booking Service API
  version: "1.0"
paths:
  /bookings/{bookingId}/cancel:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.cancelBookingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel own booking
      tags:
      - Bookings
  /bookings/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create booking payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createBookingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Bookings
  /bookings/list:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.listAllBookingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all bookings with pagination
      tags:
      - Bookings
  /bookings/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.listBookingsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List current user future bookings
      tags:
      - Bookings
  /dummyLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.dummyLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get test JWT by role
      tags:
      - Auth
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login by email/password
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.registerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register user by email/password
      tags:
      - Auth
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Create schedule payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createScheduleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create room schedule
      tags:
      - Schedules
  /rooms/{roomId}/slots/list:
    get:
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Date in YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.listSlotsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List available slots by room and date
      tags:
      - Slots
  /rooms/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create room payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createRoomResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create room
      tags:
      - Rooms
  /rooms/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.listRoomsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List rooms
      tags:
      - Rooms
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"